Move 'Feed Adapters' to separate file
